feat: 유틸함수 생성
